refactor : inputChecker -> inputConverter 로 변경
