+++ src/main/java/lotto/CalculatorApplication.java
@@ -1,9 +1,8 @@
 package lotto;
 
-import lotto.domain.Calculator;
 import lotto.domain.NumberGroup;
 import lotto.domain.OperatorGroup;
-import lotto.utility.InputChecker;
+import lotto.domain.InputConverter;
 import lotto.view.InputView;
 
 import java.util.List;
@@ -12,10 +11,10 @@ public class CalculatorApplication {
 
     public static void main(String[] args) {
         InputView inputView = new InputView();
-        InputChecker inputChecker = new InputChecker();
+        InputConverter inputConverter = new InputConverter();
 
         String formula = inputView.askFormula();
-        List<String> formulaList = inputChecker.formulaToStrList(formula);
+        List<String> formulaList = inputConverter.formulaToStrList(formula);
 
         NumberGroup numberGroup = new NumberGroup(formulaList);
         OperatorGroup operatorGroup = new OperatorGroup(formulaList);
